\name{ydhwang-package}
\alias{ydhwang-package}
\alias{ydhwang}
\docType{package}
\title{
\packageTitle{ydhwang}
}
\description{
\packageDescription{ydhwang}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ydhwang}
\packageIndices{ydhwang}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ydhwang}

Maintainer: \packageMaintainer{ydhwang}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
